Make the mask band bitmapped.  Also, remove scaling from percent slope metadata, and fix valid range for percent slope and hillshade metadata.

+++ not-validated-prototype-dswe/src/output.h
@@ -52,7 +52,7 @@ add_ps_band_product
     char *short_name,
     char *long_name,
     int min_range,
-    int max_range,
+    float max_range,
     float *data
 );
